Refactor lesson list cards
