+++ src/components/index.ts
@@ -11,6 +11,9 @@ export { ProficiencyLevelSelector } from './ProficiencyLevelSelector';
 export { LearningGoalsSelector } from './LearningGoalsSelector';
 export { DailyGoalSelector } from './DailyGoalSelector';
 export { LessonProgress } from './LessonProgress';
+export { LessonCard } from './LessonCard';
+export { TagDisplay } from './TagDisplay';
+export { TaskCardsRow } from './TaskCardsRow';
 export { TaskRenderer } from './TaskRenderer';
 export { BottomSheet } from './BottomSheet';
 export { PaywallBottomSheet } from './PaywallBottomSheet';
